refactor: relocate AI settings and simplify DI across applications
This commit standardizes the AI configuration infrastructure by moving `AiConfigurationSettings` to a central core module and removing redundant dependency injection overhead across the AshBike, KoColor, PhotoDo, and Seaweed applications.

- **Centralized Domain Models**:
    - Moved `AiConfigurationSettings` from `features:ai:configuration` to `core:data:repository`.
- **Dependency Injection Cleanup**:
    - **Removed Qualifiers**: Deleted `@Named` qualifiers (e.g., `"AshBike"`, `"KoColor"`) from `AiConfigurationSettings` and `SmartCaptureSettings` bindings in all database and data modules.
    - **Deleted Bridge Modules**: Removed `SettingsBridgeModule.kt`, `AiConfigurationBridgeModule.kt`, and `SmartCaptureBridgeModule.kt` from app modules as they are no longer required for bridging named dependencies.
    - **Simplified Injection**: Updated ViewModels in `features:analyzer`, `features:cosmetics`, `features:suggestions`, and `features:glass` to inject settings without specific qualifiers.
- **Module Dependencies**:
    - Updated `build.gradle.kts` for Seaweed's glass feature to include the `:core:data` project dependency.
