--- applications/ashbike/apps/mobile/src/main/java/com/zoewave/probase/ashbike/mobile/di/SettingsBridgeModule.kt
@@ -1,27 +1,0 @@
-package com.zoewave.probase.ashbike.mobile.di
-
-import com.zoewave.probase.features.ai.capture.domain.SmartCaptureSettings
-import com.zoewave.probase.features.ai.configuration.domain.AiConfigurationSettings
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Named
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object SettingsBridgeModule {
-
-    @Provides
-    @Singleton
-    fun provideDefaultAiSettings(
-        @Named("AshBike") impl: AiConfigurationSettings
-    ): AiConfigurationSettings = impl
-
-    @Provides
-    @Singleton
-    fun provideDefaultSmartCaptureSettings(
-        @Named("AshBike") impl: SmartCaptureSettings
-    ): SmartCaptureSettings = impl
-}
