+++ applications/ashbike/database/src/main/java/com/zoewave/probase/ashbike/database/repository/AppSettingsRepository.kt
@@ -8,7 +8,7 @@ import androidx.datastore.preferences.core.stringPreferencesKey
 import com.zoewave.ashbike.model.bike.LocationEnergyLevel
 import com.zoewave.probase.core.data.repository.SecureApiKeyRepository
 import com.zoewave.probase.features.ai.capture.domain.SmartCaptureSettings
-import com.zoewave.probase.features.ai.configuration.domain.AiConfigurationSettings
+import com.zoewave.probase.core.data.repository.AiConfigurationSettings
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
