--- applications/seaweed/apps/mobile/src/main/java/com/zoewave/probase/seaweed/mobile/di/AiConfigurationBridgeModule.kt
@@ -1,19 +1,0 @@
-package com.zoewave.probase.seaweed.mobile.di
-
-import com.zoewave.probase.features.ai.configuration.domain.AiConfigurationSettings
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object AiConfigurationBridgeModule {
-
-    @Provides
-    @Singleton
-    fun provideDefaultAiSettings(
-        impl: AiConfigurationSettings
-    ): AiConfigurationSettings = impl
-}
