--- applications/seaweed/apps/mobile/src/main/java/com/zoewave/probase/seaweed/mobile/di/SmartCaptureBridgeModule.kt
@@ -1,20 +1,0 @@
-package com.zoewave.probase.seaweed.mobile.di
-
-import com.zoewave.probase.features.ai.capture.domain.SmartCaptureSettings
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Named
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object SmartCaptureBridgeModule {
-
-    @Provides
-    @Singleton
-    fun provideDefaultSmartCaptureSettings(
-        impl: SmartCaptureSettings
-    ): SmartCaptureSettings = impl
-}
