+++ applications/seaweed/data/src/main/java/com/zoewave/probase/seaweed/data/DataModule.kt
@@ -1,7 +1,7 @@
 package com.zoewave.probase.seaweed.data
 
+import com.zoewave.probase.core.data.repository.AiConfigurationSettings
 import com.zoewave.probase.features.ai.capture.domain.SmartCaptureSettings
-import com.zoewave.probase.features.ai.configuration.domain.AiConfigurationSettings
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
