+++ applications/seaweed/data/src/main/java/com/zoewave/probase/seaweed/data/UserSettingsRepository.kt
@@ -1,7 +1,7 @@
 package com.zoewave.probase.seaweed.data
 
 import com.zoewave.probase.features.ai.capture.domain.SmartCaptureSettings
-import com.zoewave.probase.features.ai.configuration.domain.AiConfigurationSettings
+import com.zoewave.probase.core.data.repository.AiConfigurationSettings
 import com.zoewave.probase.seaweed.model.UserSettings
 import kotlinx.coroutines.flow.Flow
